BroadcastSchedulerNonCoalescing prototype
