+++ ipyparallel/client/view.py
@@ -839,16 +839,14 @@ class DirectView(View):
         ip.magics_manager.register(M)
 
 
-class BroadCastView(View):
+class BroadcastView(DirectView):
     def __init__(self, client=None, socket=None, targets=None):
         super().__init__(client=client, socket=socket, targets=targets)
 
 
     @sync_results
     @save_ids
-    def _really_apply(self, f, args=None, kwargs=None, block=None, track=None,
-                                        after=None, follow=None, timeout=None,
-                                        targets=None, retries=None):
+    def _really_apply(self, f, args=None, kwargs=None, block=None, track=None, targets=None):
         args = [] if args is None else args
         kwargs = {} if kwargs is None else kwargs
         block = self.block if block is None else block
@@ -862,14 +860,21 @@ class BroadCastView(View):
         pargs = [PrePickled(arg) for arg in args]
         pkwargs = {k: PrePickled(v) for k, v in kwargs.items()}
 
-        after = self._render_dependency(after)
-        follow = self._render_dependency(follow)
-        metadata = dict(after=after, follow=follow, timeout=timeout, targets=idents, retries=retries, is_broadcast=True)
+        metadata = dict(targets=idents, is_broadcast=True)
 
-        future = self.client.send_apply_request(
+        original_future = self.client.send_apply_request(
                 self._socket, pf, pargs, pkwargs,
                 track=track, ident=idents, metadata=metadata)
-        futures.append(future)
+
+        original_msg_id = original_future.msg_id
+
+        for ident in idents:
+            msg_and_target_id = f'{original_msg_id}_{ident.decode("utf-8")}'
+            future = self.client.create_message_futures(msg_and_target_id, async_result=True, track=True)
+            self.client.outstanding.add(msg_and_target_id)
+            self.outstanding.add(msg_and_target_id)
+            futures.append(future[0])
+
         if isinstance(targets, int):
             futures = futures[0]
         ar = AsyncResult(self.client, futures, fname=getname(f), targets=_targets,
@@ -883,43 +888,6 @@ class BroadCastView(View):
 
     def map(self, f, *sequences, **kwargs):
         pass
-
-    def _validate_dependency(self, dep):
-        """validate a dependency.
-
-        For use in `set_flags`.
-        """
-        if dep is None or isinstance(dep, string_types + (AsyncResult, Dependency)):
-            return True
-        elif isinstance(dep, (list,set, tuple)):
-            for d in dep:
-                if not isinstance(d, string_types + (AsyncResult,)):
-                    return False
-        elif isinstance(dep, dict):
-            if set(dep.keys()) != set(Dependency().as_dict().keys()):
-                return False
-            if not isinstance(dep['msg_ids'], list):
-                return False
-            for d in dep['msg_ids']:
-                if not isinstance(d, string_types):
-                    return False
-        else:
-            return False
-
-        return True
-
-    def _render_dependency(self, dep):
-        """helper for building jsonable dependencies from various input forms."""
-        if isinstance(dep, Dependency):
-            return dep.as_dict()
-        elif isinstance(dep, AsyncResult):
-            return dep.msg_ids
-        elif dep is None:
-            return []
-        else:
-            # pass to Dependency constructor
-            return list(Dependency(dep))
-
 
 
 class LoadBalancedView(View):
@@ -1249,5 +1217,5 @@ class ViewExecutor(Executor):
         if wait:
             self.view.wait()
 
-__all__ = ['LoadBalancedView', 'DirectView', 'ViewExecutor', 'BroadCastView']
+__all__ = ['LoadBalancedView', 'DirectView', 'ViewExecutor', 'BroadcastView']
